Rename Magic Item Resolver
